Removed the MB unit from the name to prevent double MB indication on y-axis of graphs
The 'MB' unit is already configured as the ValueUnit when the counter is created. Leaving the MB as a suffix on the name, causes that the 'MB' unit is shown twice on the y-axis of the graph.
Removing the 'MB' from the initial default name of the counters, will result in an improved initial usage.